+++ docs/src/changelog.md
@@ -13,6 +13,10 @@ project adheres to [Semantic Versioning](https://semver.org/).
 #### Packaging and automation
 - Add missing command-line flags to Make targets used for local development
 
+### Added
+#### Monitoring
+- Expose Prometheus metrics for periodic feed synchronization tasks
+
 ### Changed
 - Update direct and transitive dependencies